slightly lest stupid tests for simple ecdsa
