+++ src/crypto_tools/rng.rs
@@ -139,3 +139,14 @@ pub(crate) fn rng_seed_ecdsa_ephemeral_scalar(
 
     Ok(ChaCha20Rng::from_seed(seed))
 }
+
+#[cfg(test)]
+/// return the all-zero array with the first bytes set to the bytes of `index`
+pub fn dummy_secret_recovery_key(index: usize) -> SecretRecoveryKey {
+    let index_bytes = index.to_be_bytes();
+    let mut result = [0; 64];
+    for (i, &b) in index_bytes.iter().enumerate() {
+        result[i] = b;
+    }
+    SecretRecoveryKey(result)
+}
